requirements.txt
Django==3.1
django-model-utils==4.0.0
ipython==7.17.0
pre-commit==2.6.0
